api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# LangChain — install all sub-packages explicitly
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23         # ChatOpenAI pointing at NIM
langchain-huggingface==0.0.3     # HuggingFaceEmbeddings (moved out of community)

# HuggingFace + embeddings
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.2.2                     # CPU build — swap for torch+cu121 on GPU host

# Vector store
faiss-cpu==1.8.0                 # swap to faiss-gpu on CUDA host

# BM25 sparse retrieval
rank-bm25==0.2.2

# HTTP
httpx==0.27.0
python-multipart==0.0.9

# Env vars
python-dotenv==1.0.1

# Etsy API connector
httpx==0.27.0                    # already listed above — async HTTP client
